IKASAN-1865 - Move TransferManagerProgressLogger.java so its in the correct package with AmazonS3Client.java
